--- ecom/src/main/java/com/ecom/ecom/Repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.ecom.ecom.repository;
-
-import com.example.demo.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
