--- ecom/src/main/java/com/ecom/ecom/Repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.ecom.ecom.Repository;
-
-import com.ecom.ecom.Entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
