--- ecom/src/main/java/com/ecom/ecom/Services/CategoryService.java
@@ -1,33 +1,0 @@
-package com.ecom.ecom.service;
-
-import com.ecom.ecom.entity.Category;
-import com.ecom.ecom.repository.CategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class CategoryService {
-
-    @Autowired
-    private CategoryRepository categoryRepository;
-
-    public Page<Category> getAllCategories(Pageable pageable) {
-        return categoryRepository.findAll(pageable);
-    }
-
-    public Optional<Category> getCategoryById(Long id) {
-        return categoryRepository.findById(id);
-    }
-
-    public Category saveCategory(Category category) {
-        return categoryRepository.save(category);
-    }
-
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
-    }
-}
