+++ ecom/src/main/java/com/ecom/ecom/repository/CategoryRepository.java
@@ -1,0 +1,9 @@
+package com.ecom.ecom.repository;
+
+import com.ecom.ecom.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
