+++ ecom/src/main/java/com/ecom/ecom/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.ecom.ecom.repository;
+
+import com.ecom.ecom.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
